[Tech] changes after design review
