+++ frontend/src/features/Regulation/components/RegulationForm/custom_form/CustomSelectComponent.tsx
@@ -1,5 +1,4 @@
 import { Select } from '@mtes-mct/monitor-ui'
-import styled from 'styled-components'
 
 import type { ReactNode } from 'react'
 
@@ -36,36 +35,28 @@ export function CustomSelectComponent({
   width = undefined
 }: CustomSelectComponentProps) {
   return (
-    <SelectWrapper>
-      <Select
-        cleanable={cleanable}
-        data-cy={dataCy}
-        disabled={disabled}
-        isLabelHidden
-        isTransparent
-        label={label}
-        locale={
-          {
-            emptyMessage,
-            noResultsText: emptyMessage
-          } as any
-        }
-        name={name}
-        onChange={onChange}
-        options={options}
-        placeholder={placeholder}
-        renderMenuItem={renderMenuItem}
-        searchable={searchable}
-        style={{ width: width ? `${width}px` : '200px' }}
-        value={value}
-        virtualized={virtualized}
-      />
-    </SelectWrapper>
+    <Select
+      cleanable={cleanable}
+      data-cy={dataCy}
+      disabled={disabled}
+      isLabelHidden
+      isTransparent
+      label={label}
+      locale={
+        {
+          emptyMessage,
+          noResultsText: emptyMessage
+        } as any
+      }
+      name={name}
+      onChange={onChange}
+      options={options}
+      placeholder={placeholder}
+      renderMenuItem={renderMenuItem}
+      searchable={searchable}
+      style={{ width: width ? `${width}px` : '200px' }}
+      value={value}
+      virtualized={virtualized}
+    />
   )
 }
-
-const SelectWrapper = styled.div`
-  display: inline-block;
-  margin: 0px 10px 0px 0px;
-  vertical-align: sub;
-`
