+++ frontend/src/features/Regulation/components/RegulationTables/tableCells.tsx
@@ -1,7 +1,7 @@
 // TODO Re-enable ESLint for this old file which has been migrated from JSX to TSX.
 /* eslint-disable */
 
-import { Accent, Icon, IconButton, Select, Tag, THEME } from '@mtes-mct/monitor-ui'
+import { Accent, Icon, IconButton, Tag, THEME } from '@mtes-mct/monitor-ui'
 import { useCallback, type ChangeEvent } from 'react'
 import { Table } from 'rsuite'
 import styled from 'styled-components'
@@ -140,43 +140,6 @@ export function FleetSegmentInput({
     />
   )
 }
-
-export function ControlPriorityCell({ dataKey, onChange, ...props }) {
-  const { rowData } = props
-  const dataCy = `row-${rowData.id}-${dataKey}`
-
-  return (
-    <StyledCell key={rowData.id} {...props} className="table-content-editing">
-      <Select
-        container={document.body}
-        cleanable={false}
-        isTransparent
-        isLabelHidden
-        label="Priorité de contrôle"
-        name="controlPriority"
-        options={[
-          { label: '1', value: 1 },
-          { label: '2', value: 2 },
-          { label: '3', value: 3 },
-          { label: '4', value: 4 }
-        ]}
-        data-cy={dataCy}
-        onChange={value => {
-          const controlPriority = value ?? ''
-          onChange && onChange(rowData.id, dataKey, controlPriority)
-        }}
-        searchable={false}
-        value={rowData[dataKey]}
-      />
-    </StyledCell>
-  )
-}
-
-const StyledCell = styled(Cell)`
-  > div {
-    padding: 4px 0 0 0 !important;
-  }
-`
 
 type SegmentCellWithTitleProps = Readonly<
   Omit<InnerCellProps<any, any>, 'rowData'> &
