+++ frontend/src/features/commonStyles/Input.style.jsx
@@ -1,6 +1,7 @@
-import styled from 'styled-components'
-import { COLORS } from '../../constants/constants'
 import { Input } from 'rsuite'
+import styled from 'styled-components'
+
+import { COLORS } from '../../constants/constants'
 
 export const Label = styled.span`
   text-align: left;
@@ -8,18 +9,18 @@ export const Label = styled.span`
   min-width: 150px;
   display: inline-block;
   font-size: 13px;
-  ${props => props.isLast ? '' : 'margin-right: 20px'};
+  ${props => (props.isLast ? '' : 'margin-right: 20px')};
 `
 
 export const CustomInput = styled(Input)`
   font-size: 13px;
-  ${props => props.width ? '' : 'min-width: 100px;'}
-  ${props => props.width ? `width: ${props.width};` : ''}
-  border: 1px solid ${props => props.$isRed ? `${COLORS.maximumRed}` : `${COLORS.lightGray}`};
+  ${props => (props.width ? '' : 'min-width: 100px;')}
+  ${props => (props.width ? `width: ${props.width};` : '')}
+  border: 1px solid ${props => (props.$isRed ? `${COLORS.maximumRed}` : `${COLORS.lightGray}`)};
   border-radius: 2px;
   color: ${COLORS.gunMetal}!important;
   font-weight: 500;
-  background-color: ${props => props.$isGray ? COLORS.gainsboro : COLORS.white};
+  background-color: ${props => (props.$isGray ? COLORS.gainsboro : COLORS.white)};
   margin: 0px 10px 0px 0px;
   padding: 8px;
